requirements.txt
# VN Market Intelligence - Agent 1, demo, dashboard
# Pin numpy<2 for compatibility with older matplotlib/tensorflow stacks
numpy>=1.26,<2
anthropic>=0.40.0
google-generativeai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
PyYAML>=6.0
pandas>=2.0.0
# Technical indicators: plain pandas in technical_calc.py (no pandas-ta)
vnstock>=3.0.0
feedparser>=6.0.0
requests>=2.31.0
streamlit>=1.39.0
plotly>=5.24.0
